[AG-17118] Refactor(pagination): rename module file and link JsDoc to page numbers panel
