+++ packages/ag-grid-community/src/allCommunityModule.ts
@@ -31,8 +31,8 @@ import type { _ModuleWithoutApi } from './interfaces/iModule';
 import { EventApiModule } from './misc/apiEvents/apiEventModule';
 import { LocaleModule } from './misc/locale/localeModule';
 import { GridStateModule } from './misc/state/stateModule';
-import { PaginationPageNumbersModule } from './pagination/pageNumbersModule';
 import { PaginationModule } from './pagination/paginationModule';
+import { PaginationPageNumbersModule } from './pagination/paginationPageNumbersModule';
 import { PinnedRowModule } from './pinnedRowModel/pinnedRowModule';
 import { HighlightChangesModule } from './rendering/cell/highlightChangesModule';
 import { RenderApiModule } from './rendering/renderModule';
